backend/pyproject.toml
[project]
name = "meld-backend"
version = "0.1.0"
description = "Meld Health Coach, Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    # Sync driver for Alembic. asyncpg is async-only; alembic's command.upgrade
    # uses the bare `postgresql://` URL converted in app/database.py and needs
    # a sync DBAPI. Without this, the migration thread silently hangs trying to
    # resolve a non-existent default driver. See incident 2026-04-29.
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "anthropic>=0.49.0",
    "httpx[http2]>=0.28.0",
    "pydantic-settings>=2.7.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.21.0",
    "pillow>=11.0.0",
    "garminconnect>=0.2.0",
    "pylotoncycle>=0.9.1",
    "textstat>=0.7.0",
    "pyjwt[crypto]>=2.10.1",
    "slowapi>=0.1.9",
    "email-validator>=2.2.0",
    "sentry-sdk[fastapi]>=2.0.0",
    # ── Signal Engine ML stack (see ~/.claude/plans/golden-floating-creek.md)
    # These are imported LAZILY inside backend/ml/api.py function bodies, never
    # at module import time. The boundary test (tests/ml/test_boundary.py) and
    # cold-boot test (tests/ml/test_cold_boot.py) enforce this invariant. If you
    # add a top-level import of any of these anywhere under backend/app/, the
    # FastAPI cold boot budget (< 4s on Railway) will be blown.
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "ruptures>=1.1.9",
    "pycatch22>=0.4.5",
    "shap>=0.46.0",
    "hdbscan>=0.8.38",
    "mlflow>=2.15.0",
    "evidently>=0.4.0",
    "nannyml>=0.12.0",
    # Heavier deps, Modal-only in practice but installable on Railway for parity.
    # Prophet pulls cmdstanpy; dowhy pulls networkx + graphviz; coremltools is
    # macOS/linux-only and ~150MB. Keep them here so `uv sync` works everywhere.
    "prophet>=1.1.5",
    "dowhy>=0.11.0",
    "econml>=0.15.0",
    "boto3>=1.35.0",
    # Direct CVE pins on transitive deps. See docs/comprehensive-scan-2026-04-30.md
    # and the per-CVE notes below for context.
    "cryptography>=46.0.7",  # CVE-2026-39892: buffer overflow on non-contig buffers
    "mako>=1.3.11",          # path traversal via // URI in TemplateLookup
    # NOTE: lightgbm CVE-2024-43598 (RCE via crafted model file) cannot be patched
    # right now: nannyml 0.13.1 (latest) caps lightgbm<4.6. Our usage is
    # transitive only (no direct lightgbm imports anywhere in backend/), so the
    # CVE has no exploit path in current code. Tracked as a separate Tier 2
    # decision (drop nannyml or wait for upstream).
]

[project.optional-dependencies]
coreml = ["coremltools>=8.0"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
# Resolve only for the two envs we actually ship to: Brock's macOS arm64 dev
# laptop and the Linux x86_64 Railway container. Without this, uv tries to
# satisfy every possible platform and falls over on transitive deps like
# kaleido 0.2.x whose old wheels predate macOS arm64.
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]
